% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bivalves}
\alias{bivalves}
\title{Bivalve dataset}
\format{
a tibble with 8,652 rows and 7 variables
\describe{
  \item{Date}{Sample collection date.}
  \item{Station}{Station where sample was collected.}
  \item{Taxa}{Bivalve species name.}
  \item{CPUE}{Catch per unit effort in number of clams (\ifelse{html}{\out{m<sup>-2</sup>}}{\eqn{m^{-2}}}).}
  \item{Year}{Year sample was collected.}
  \item{MonthYear}{Month and year of sample collection.}
  \item{Source}{Name of the source dataset.}
  }
}
\usage{
bivalves
}
\description{
Bivalve abundance dataset from the California Department of Water Resources Environmental Monitoring Program.
}
\details{
More metadata and information on methods are available \href{https://emp.baydeltalive.com/projects/11280}{here}.
}
\seealso{
\code{\link{DeltaBivalver}}, \code{\link{DeltaDater}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
