% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dayflow}
\alias{dayflow}
\title{Dayflow dataset}
\format{a tibble with 8,652 rows and 7 variables
\describe{
  \item{Date}{Date.}
  \item{Out}{Delta outflow (\eqn{ft^{3}} \eqn{s^{-1}}.)}
  \item{X2}{X2 (km).}
  \item{CPUE}{Catch per unit effort in number of clams \eqn{m^{-2}}.}
  \item{Year}{Year sample was collected.}
  \item{MonthYear}{Month and year of sample collection.}
  \item{Source}{Name of the source dataset.}
  }}
\usage{
dayflow
}
\description{
Outflow and X2 from the California Department of Water Resources Dayflow model.
}
\details{
More metadata and information on methods are available \href{https://data.cnra.ca.gov/dataset/dayflow}{here}.
}
\seealso{
\code{\link{DSCBivalver}}, \code{\link{DSCDater}}, \code{\link{DSCMetadater}}
}
\keyword{datasets}
