% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaDater.R
\name{DeltaDater}
\alias{DeltaDater}
\title{Process report data}
\usage{
DeltaDater(
  Start_year = 2002,
  End_year = 2020,
  Variables = c("Bivalves", "Zooplankton", "Phytoplankton", "Water quality"),
  WQ_sources = c("EMP", "STN", "FMWT", "EDSM", "SKT", "20mm", "Suisun"),
  Shapefile = deltamapr::R_EDSM_Strata_1819P1,
  Region_column = "Stratum",
  Regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Sac Deep Water Shipping Channel", "Cache Slough/Liberty Island",
    "Lower Joaquin River", "Southern Delta"),
  Phyt_start = 2008
)
}
\arguments{
\item{Start_year}{Earliest year you would like included in the report. Must be an integer. Defaults to \code{2002}. For these purposes, December is pushed to the next year to ensure Winters are in the same year.}

\item{End_year}{Latest year you would like included in the dataset. Must be an integer. Defaults to \code{2020}. For these purposes, December is pushed to the next year to ensure Winters are in the same year.}

\item{Variables}{Character vector of variables you would like included in the dataset.
Defaults to all possible options: \code{Variables = c("Bivalves", "Zooplankton", "Phytoplankton", "Water quality")}.}

\item{WQ_sources}{Character vector of data sources for the water quality variables, pulled from the \code{\link[discretewq]{discretewq}} package.
See \code{\link[discretewq]{wq}} for choices.}

\item{Shapefile}{Shapefile you would like used to define regions in the dataset. Must be in \code{\link[sf]{sf}} format, e.g., imported with \code{\link[sf]{st_read}}. Defaults to \code{\link[deltamapr]{R_EDSM_Strata_1819P1}}.}

\item{Region_column}{Quoted name of the column in the Shapefile with the region designations.}

\item{Regions}{Character vector of regions to be included in the dataset. Must correspond with levels of the \code{Region_column}. To include all data points regardless of whether they correspond to a region in the \code{Shapefile} set \code{Regions = NULL}.}

\item{Phyt_start}{First year to include for phytoplankton data. Defaults to \code{2008}, when better counting methods were first adopted. This parameter uses calendar year.}
}
\value{
A list of datasets
}
\description{
Imports, filters, and processes datasets and outputs a list of desired datasets
}
\examples{
Data <- DeltaDater(Start_year = 1900,
WQ_sources = c("EMP", "STN", "FMWT", "EDSM", "SKT", "20mm", "Suisun"),
Variables = "Water quality",
Regions = NULL)
}
