% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaMapper.R
\name{DeltaMapper}
\alias{DeltaMapper}
\title{Plot Delta Smelt data}
\usage{
DeltaMapper(
  Regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Sac Deep Water Shipping Channel", "Cache Slough/Liberty Island",
    "Lower Joaquin River", "Southern Delta"),
  Save = FALSE,
  Save_location
)
}
\arguments{
\item{Regions}{Character vector of regions to include in the plot. The data will be filtered to only include these regions and ordered in the order provided here. To include data with NA regions, set \code{Regions=NULL}.}

\item{Save}{Logical. Should plot be saved?}

\item{Save_location}{If \code{Save=TRUE}, where should plot be saved? Must end with a filename with a ".png" extension.}
}
\value{
A list with the plot and processed data.
}
\description{
Function to process and plot Delta Smelt data
}
\seealso{
\code{\link{DeltaDater}}
}
