% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deltaregions}
\alias{deltaregions}
\title{Delta regions}
\format{
an sf tibble with 8 rows and 2 variables
\describe{
  \item{Stratum}{Region.}
  \item{geometry}{Polygon coordinates.}
  }
}
\usage{
deltaregions
}
\description{
Region polygons for the Sacramento San Joaquin Delta from the United States Fish and Wildlife Service Enhanced Delta Smelt Monitoring Program.
}
\seealso{
\code{\link{DSCMapper}}, \code{\link{DSCDater}}
}
\keyword{datasets}
