% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltareportr.R
\docType{package}
\name{deltareportr}
\alias{deltareportr}
\title{deltareportr: A package to create automated reports for the Sacramento San Joaquin Delta}
\description{
This package contains a number of functions and datasets to generate automated reports on the Delta, as well as the reports themselves in the analysis folder.
}
