% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaSmelter.R
\name{DeltaSmelter}
\alias{DeltaSmelter}
\title{Plot Delta Smelt data}
\usage{
DeltaSmelter(
  Start_year = 2002,
  End_year = 2018,
  EDSM_regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Sac Deep Water Shipping Channel", "Cache Slough/Liberty Island",
    "Lower Joaquin River")
)
}
\arguments{
\item{Start_year}{First year (integer) that should be included in the plot}

\item{End_year}{Last year (integer) that should be included in the plot. This year will also be highlighted.}

\item{EDSM_regions}{Character vector of regions to include in the EDSM plot. The data will be filtered to only include these regions and ordered in the order provided here. To include data with NA regions, set \code{Regions=NULL}. No Delta Smelt were caught in the Southern Delta so that region is excluded by default.}
}
\value{
A list with the plot and processed data.
}
\description{
Function to process and plot Delta Smelt data
}
\seealso{
\code{\link{DeltaDater}}
}
