% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaSmeltReporter.R
\name{DeltaSmeltReporter}
\alias{DeltaSmeltReporter}
\title{Produce Delta Smelt Conditions Report}
\usage{
DeltaSmeltReporter(File, ...)
}
\arguments{
\item{File}{Filename and path to location where the report should be saved.}

\item{...}{Additional parameters to provide to \code{\link[rmarkdown]{render}}.}
}
\description{
Function to render the Delta Smelt Conditions Report
}
\details{
See \code{\link[rmarkdown]{render}} for more details.
}
\keyword{internal}
