% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaWQer.R
\name{DeltaWQer}
\alias{DeltaWQer}
\title{Plot water quality data}
\usage{
DeltaWQer(
  Data,
  End_year,
  Start_year = 2002,
  Regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Sac Deep Water Shipping Channel", "Cache Slough/Liberty Island",
    "Lower Joaquin River", "Southern Delta"),
  Temp_seasons = c("Winter", "Spring", "Summer", "Fall"),
  Secchi_seasons = c("Winter", "Spring", "Summer", "Fall"),
  Salinity_seasons = c("Winter", "Spring", "Summer", "Fall"),
  Chl_seasons = c("Winter", "Spring", "Summer", "Fall"),
  Micro_seasons = c("Winter", "Spring", "Summer", "Fall")
)
}
\arguments{
\item{Data}{Input dataset created by \code{\link{DeltaDater}}.}

\item{End_year}{Last year (integer) that should be included in the plot. This year will also be highlighted.}

\item{Start_year}{First year (integer) that should be included in the plot}

\item{Regions}{Character vector of regions to include in the plot. The data will be filtered to only include these regions and ordered in the order provided here. To include data with NA regions, set \code{Regions=NULL}.}

\item{Temp_seasons}{Character vector of seasons to retain for temperature. One plot will be produced for each season. Should be a combination of "Summer", "Fall", "Winter", or "Spring".}

\item{Secchi_seasons}{Character vector of seasons to retain for Secchi. One plot will be produced for each season. Should be a combination of "Summer", "Fall", "Winter", or "Spring".}

\item{Salinity_seasons}{Character vector of seasons to retain for salinity. One plot will be produced for each season. Should be a combination of "Summer", "Fall", "Winter", or "Spring".}

\item{Chl_seasons}{Character vector of seasons to retain for chlorophyll. One plot will be produced for each season. Should be a combination of "Summer", "Fall", "Winter", or "Spring".}

\item{Micro_seasons}{Character vector of seasons to retain for \emph{Microcystis}. One plot will be produced for each season. Should be a combination of "Summer", "Fall", "Winter", or "Spring".}
}
\value{
A list with the plot and processed data.
}
\description{
Function to process and plot water quality (temperature, Secchi depth, salinity, chlorophyll, and \emph{Microcystis}) data
}
\seealso{
\code{\link{DeltaDater}}
}
