% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSCDater.R
\name{DSCDater}
\alias{DSCDater}
\title{Process report data}
\usage{
DSCDater(
  Start_year = 2002,
  Variables = c("Bivalves", "Zooplankton", "Phytoplankton", "Water quality"),
  Shapefile = deltareportr::deltaregions,
  Region_column = "Stratum",
  Regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Sac Deep Water Shipping Channel", "Cache Slough/Liberty Island",
    "Lower Joaquin River", "Southern Delta")
)
}
\arguments{
\item{Start_year}{Earliest year you would like included in the report. Must be an integer. Defaults to \code{2002}.}

\item{Variables}{Character vector of variables you would like included in the dataset. Defaults to all possible options: \code{Variables = c("Bivalves", "Zooplankton", "Phytoplankton", "Water quality")}.}

\item{Shapefile}{Shapefile you would like used to define regions in the dataset. Must be in \code{\link[sf]{sf}} format, e.g., imported with \code{\link[sf]{st_read}}. Defaults to \code{\link{deltaregions}}.}

\item{Region_column}{Quoted name of the column in the Shapefile with the region designations.}

\item{Regions}{Character vector of regions to be included in the dataset. Must correspond with levels of the \code{Region_column}. To include all data points regardless of whether they correspond to a region in the \code{Shapefile} set \code{Regions = NULL}.}
}
\value{
A list of datasets
}
\description{
Imports, filters, and processes datasets and outputs a list of desired datasets
}
