% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSCMetadater.R
\name{DSCMetadater}
\alias{DSCMetadater}
\title{Plot Metadata}
\usage{
DSCMetadater(
  Data,
  Start_year = 2002,
  Regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Sac Deep Water Shipping Channel", "Cache Slough/Liberty Island",
    "Lower Joaquin River", "Southern Delta")
)
}
\arguments{
\item{Data}{Input dataset created by \code{\link{DSCDater}}.}

\item{Start_year}{First year (integer) that should be included in the plot}

\item{Regions}{Character vector of regions to include in the plot. The data will be filtered to only include these regions and ordered in the order provided here. To include data with NA regions, set \code{Regions=NULL}.}
}
\value{
A list with the plot and processed data.
}
\description{
Function to process and plot metadata
}
\seealso{
\code{\link{DSCDater}}
}
