% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSCZooper.R
\name{DSCZooper}
\alias{DSCZooper}
\title{Plot phytoplankton data}
\usage{
DSCZooper(
  Data,
  Start_year = 2002,
  End_year = 2018,
  Regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Lower Joaquin River", "Southern Delta"),
  Seasons = "Fall"
)
}
\arguments{
\item{Data}{Input dataset created by \code{\link{DSCDater}}.}

\item{Start_year}{First year (integer) that should be included in the plot}

\item{End_year}{Last year (integer) that should be included in the plot. This year will also be highlighted.}

\item{Regions}{Character vector of regions to include in the plot. The data will be filtered to only include these regions and ordered in the order provided here. To include data with NA regions, set \code{Regions=NULL}.}

\item{Seasons}{Character vector of seasons to include in the dataset. Should be one of "Summer", "Fall", "Winter", or "Spring".}
}
\value{
A list with the plot and processed data.
}
\description{
Function to process and plot phytoplankton data
}
\seealso{
\code{\link{DSCDater}}
}
