% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_minor.R
\name{insert_minor}
\alias{insert_minor}
\title{Minor axis breaks}
\usage{
insert_minor(major_labs, n_minor)
}
\arguments{
\item{major_labs}{Vector with sequence of major labels}

\item{n_minor}{Number of minor breaks to place between major breaks.}
}
\description{
Function to create minor axis breaks
}
\keyword{internal}
