% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phyto}
\alias{phyto}
\title{Phytoplankton dataset}
\format{a tibble with 8,652 rows and 7 variables
\describe{
  \item{Date}{Sample collection date.}
  \item{Station}{Station where sample was collected.}
  \item{Taxa}{Phytoplankton taxa.}
  \item{CPUE}{Catch per unit effort in number of cells, colonies, or filaments \eqn{ml^{-1}}.}
  \item{Year}{Year sample was collected.}
  \item{MonthYear}{Month and year of sample collection.}
  \item{Source}{Name of the source dataset.}
  }}
\usage{
phyto
}
\description{
Phytoplankton abundance dataset from the California Department of Water Resources Environmental Monitoring Program.
}
\details{
More metadata and information on methods are available \href{https://emp.baydeltalive.com/projects/11282}{here}.
}
\seealso{
\code{\link{DSCPhyter}}, \code{\link{DSCDater}}, \code{\link{DSCMetadater}}
}
\keyword{datasets}
