% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smelt_iep}
\alias{smelt_iep}
\title{Delta Smelt IEP indices}
\format{
a tibble with 157 rows and 3 variables
\describe{
  \item{Year}{Index year.}
  \item{Index}{Delta Smelt index number.}
  \item{Source}{Name of the source dataset.}
  }
}
\usage{
smelt_iep
}
\description{
Delta Smelt abundance indices from the California Department of Fish and Wildlife surveys.
}
\details{
More metadata and information on methods are available here: \href{https://www.dfg.ca.gov/delta/projects.asp?ProjectID=FMWT}{Fall Midwater Trawl}, \href{https://wildlife.ca.gov/Conservation/Delta/Spring-Kodiak-Trawl}{Spring Kodiak Trawl}, \href{https://wildlife.ca.gov/Conservation/Delta/20mm-Survey}{20mm Survey}, and \href{https://wildlife.ca.gov/Conservation/Delta/Townet-Survey}{Summer Townet}.
}
\seealso{
\code{\link{DeltaSmelter}}
}
\keyword{datasets}
