% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationBrowser.R
\name{StationBrowser}
\alias{StationBrowser}
\title{Browse station locations}
\usage{
StationBrowser(
  Data = deltareportr::stations,
  Sources = NULL,
  StationIDs = NULL
)
}
\arguments{
\item{Data}{Station location data. Defaults to \code{\link{stations}}, which only incldues fixed station locations + EMP EZ locations post 2004. Does not incldue EDSM.
Data must include columns named Latitude, Longitude, Source, and StationID.}

\item{Sources}{Which data sources would you like to include? Set \code{Sources=NULL} (the default) to include all.}

\item{StationIDs}{Which StationIDs to include. StationIDs take the form "Source, Station". Set \code{StationIDs=NULL} (the default) to include all.}
}
\description{
Launches a leaflet map to browse station locations.
}
