% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_20mm}
\alias{wq_20mm}
\title{20mm water quality data}
\format{
a tibble with 9,458 rows and 12 variables
\describe{
  \item{Station}{Station where sample was collected.}
  \item{Temperature}{Temperature in °C.}
  \item{Conductivity}{Specific conductance (\eqn{\mu}S \eqn{cm^{-1}}) at surface.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Notes}{Comments.}
  \item{Latitude}{Latitude in decimal degrees.}
  \item{Longitude}{Longitude in decimal degrees.}
  \item{Date}{Date sample was collected.}
  \item{Tide}{Tidal stage.}
  \item{Depth}{Bottom depth (m).}
  \item{Datetime}{Date and time of sample collection.}
  \item{Source}{Name of the source dataset.}
  }
}
\usage{
wq_20mm
}
\description{
Water quality data from the California Department of Fish and Wildlife 20mm survey.
}
\details{
More metadata and information on methods are available \href{https://wildlife.ca.gov/Conservation/Delta/20mm-Survey}{here}.
}
\seealso{
\code{\link{DSCDater}}, \code{\link{DSCWQer}}, \code{\link{DSCMetadater}}
}
\keyword{datasets}
