% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_djfmp}
\alias{wq_djfmp}
\title{DJFMP water quality data}
\format{
a tibble with 136,599 rows and 6 variables
\describe{
  \item{Station}{Station where sample was collected.}
  \item{Date}{Date sample was collected.}
  \item{Temperature}{Temperature in °C.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Source}{Name of the source dataset.}
  \item{Datetime}{Date and time of sample collection.}
  }
}
\usage{
wq_djfmp
}
\description{
Water quality data from the United States Fish and Wildlife Service Delta Juvenile Fish Monitoring Program.
}
\details{
More metadata and information on methods are available \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.244.3}{here}.
}
\seealso{
\code{\link{DeltaDater}}, \code{\link{DeltaWQer}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
