% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_edsm}
\alias{wq_edsm}
\title{EDSM water quality data}
\format{a tibble with 14,843 rows and 7 variables
\describe{
  \item{Date}{Date sample was collected.}
  \item{Latitude}{Latitude in decimal degrees.}
  \item{Longitude}{Longitude in decimal degrees.}
  \item{Temperature}{Temperature in °C.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Station}{Station where sample was collected.}
  \item{Source}{Name of the source dataset.}
  }}
\usage{
wq_edsm
}
\description{
Water quality data from the United States Fish and Wildlife Service Enhanced Delta Smelt Monitoring Program.
}
\details{
More metadata and information on methods are available \href{https://www.fws.gov/lodi/juvenile_fish_monitoring_program/jfmp_index.htm}{here}.
}
\seealso{
\code{\link{DSCDater}}, \code{\link{DSCWQer}}, \code{\link{DSCMetadater}}
}
\keyword{datasets}
