% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_emp}
\alias{wq_emp}
\title{EMP water quality data}
\format{
a tibble with 16,293 rows and 8 variables
\describe{
  \item{Date}{Date sample was collected.}
  \item{Station}{Station where sample was collected.}
  \item{Conductivity}{Specific conductance (\eqn{\mu}S \eqn{cm^{-1}}) at surface.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Temperature}{Temperature (°C) at surface.}
  \item{Chlorophyll}{Chlorophyll concentration (\eqn{\mu}g \eqn{L^{-1}})}
  \item{Microcystis}{Microcystis bloom intensity on a qualitative scale from 1 to 5 where 1 = absent, 2 = low, 3 = medium, 4 = high, and 5 = very high.}
  \item{Source}{Name of the source dataset.}
  }
}
\usage{
wq_emp
}
\description{
Water quality data from the California Department of Water Resources Environmental Monitoring Program.
}
\details{
More metadata and information on methods are available \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.458.2}{here}.
}
\seealso{
\code{\link{DSCDater}}, \code{\link{DSCWQer}}, \code{\link{DSCMetadater}}
}
\keyword{datasets}
