% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_skt}
\alias{wq_skt}
\title{SKT water quality data}
\format{
a tibble with 3,965 rows and 12 variables
\describe{
  \item{Date}{Date sample was collected.}
  \item{Station}{Station where sample was collected.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Conductivity}{Specific conductance (\eqn{\mu}S \eqn{cm^{-1}}) at surface.}
  \item{Temperature}{Temperature (°C) at surface.}
  \item{Depth}{Bottom depth in ???}
  \item{Tide}{Tidal stage.}
  \item{Notes}{Comments.}
  \item{Source}{Name of the source dataset.}
  \item{Latitude}{Latitude in decimal degrees.}
  \item{Longitude}{Longitude in decimal degrees.}
  \item{Datetime}{Date and time of sample collection.}
  }
}
\usage{
wq_skt
}
\description{
Water quality data from the California Department of Fish and Wildlife Spring Kodiak Trawl.
}
\details{
More metadata and information on methods are available \href{http://www.dfg.ca.gov/delta/projects.asp?ProjectID=SKT}{here}.
}
\seealso{
\code{\link{DSCDater}}, \code{\link{DSCWQer}}, \code{\link{DSCMetadater}}
}
\keyword{datasets}
