% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_stn}
\alias{wq_stn}
\title{STN water quality data}
\format{a tibble with 8,266 rows and 7 variables
\describe{
  \item{Date}{Date sample was collected.}
  \item{Station}{Station where sample was collected.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Temperature}{Temperature (°C) at surface.}
  \item{Conductivity}{Specific conductance (\eqn{\mu}S \eqn{cm^{-1}}) at surface.}
  \item{Microcystis}{Microcystis bloom intensity on a qualitative scale from 1 to 5 where 1 = absent, 2 = low, 3 = medium, 4 = high, and 5 = very high.}
  \item{Source}{Name of the source dataset.}
  }}
\usage{
wq_stn
}
\description{
Water quality data from the California Department of Fish and Wildlife Summer Townet.
}
\details{
More metadata and information on methods are available \href{https://wildlife.ca.gov/Conservation/Delta/Townet-Survey}{here}.
}
\seealso{
\code{\link{DSCDater}}, \code{\link{DSCWQer}}, \code{\link{DSCMetadater}}
}
\keyword{datasets}
