% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_suisun}
\alias{wq_suisun}
\title{Suisun water quality data}
\format{
a tibble with 14,253 rows and 10 variables
\describe{
  \item{Station}{Station where sample was collected.}
  \item{Date}{Date sample was collected.}
  \item{QADone}{Was QA completed?}
  \item{Temperature}{Temperature (°C) at surface.}
  \item{Salinity}{Salinity (ppt) at the surface.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Conductivity}{Specific conductance (\eqn{\mu}S \eqn{cm^{-1}}) at surface.}
  \item{Tide}{Tidal stage.}
  \item{Datetime}{Date and time of sample collection.}
  \item{Source}{Name of the source dataset.}
  \item{Depth}{Bottom depth (m).}
  }
}
\usage{
wq_suisun
}
\description{
Water quality data from the UC Davis Suisun Marsh Fish Study.
}
\details{
More metadata and information on methods are available \href{https://watershed.ucdavis.edu/project/suisun-marsh-fish-study}{here}.
}
\seealso{
\code{\link{DeltaDater}}, \code{\link{DeltaWQer}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
