% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoop_mass_conversions}
\alias{zoop_mass_conversions}
\title{Zooplankton mass conversions}
\format{
a tibble with 40 rows and 2 variables
\describe{
  \item{Mass}{Average individual mass (\eqn{\mu}g).}
  \item{Taxlifestage}{Taxonomic name and lifestage.}
  }
}
\usage{
zoop_mass_conversions
}
\description{
Average dry mass for each meso and micro zooplankton taxa
}
\seealso{
\code{\link{DeltaZooper}}, \code{\link{DeltaDater}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
