% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoop_mysid}
\alias{zoop_mysid}
\title{Mysid data}
\format{
a tibble with 40 rows and 2 variables
\describe{
  \item{Date}{Date sample was collected.}
  \item{Station}{Station where sample was collected.}
  \item{Taxa}{Taxonomic name.}
  \item{BPUE}{Biomass per unit effort (\eqn{\mu}g \eqn{m^{-3}}).}
  \item{Year}{Year sample was collected.}
  \item{MonthYear}{Month and year sample was collected.}
  \item{Source}{Name of the source dataset.}
  }
}
\usage{
zoop_mysid
}
\description{
Mysid biomass per unit effort from the California Department of Fish and Wildlife Environmental Monitoring Program
}
\details{
More metadata and information on methods are available \href{https://wildlife.ca.gov/Conservation/Delta/Zooplankton-Study}{here}.
}
\seealso{
\code{\link{DeltaZooper}}, \code{\link{DeltaDater}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
