% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dayflow}
\alias{dayflow}
\title{Dayflow dataset}
\format{
a tibble with 8,652 rows and 7 variables
\describe{
  \item{Date}{Date.}
  \item{Out}{Delta outflow (\ifelse{html}{\out{ft<sup>3</sup>s<sup>-1</sup>}}{\eqn{ft^{3}} \eqn{s^{-1}}}).}
  \item{X2}{X2 (km).}
  }
}
\usage{
dayflow
}
\description{
Outflow and X2 from the California Department of Water Resources Dayflow model.
}
\details{
More metadata and information on methods are available \href{https://data.cnra.ca.gov/dataset/dayflow}{here}.
}
\seealso{
\code{\link{DeltaBivalver}}, \code{\link{DeltaDater}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
