% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaDater.R
\name{DeltaDater}
\alias{DeltaDater}
\title{Process report data}
\usage{
DeltaDater(
  Start_year = 2002,
  Variables = c("Bivalves", "Zooplankton", "Phytoplankton", "Water quality"),
  WQ_sources = c("EMP", "STN", "FMWT", "EDSM"),
  Shapefile = deltareportr::deltaregions,
  Region_column = "Stratum",
  Regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Sac Deep Water Shipping Channel", "Cache Slough/Liberty Island",
    "Lower Joaquin River", "Southern Delta")
)
}
\arguments{
\item{Start_year}{Earliest year you would like included in the report. Must be an integer. Defaults to \code{2002}.}

\item{Variables}{Character vector of variables you would like included in the dataset. Defaults to all possible options: \code{Variables = c("Bivalves", "Zooplankton", "Phytoplankton", "Water quality")}.}

\item{WQ_sources}{Character vector of data sources for the water quality variables. Choices include "EMP" (Environmental Monitoring Program, \code{\link{wq_emp}}), "STN" (Summer Townet Survey, \code{\link{wq_stn}}), "FMWT" (Fall Midwater Trawl, \code{\link{wq_fmwt}}), "EDSM" (Enhanced Delta Smelt Monitoring, \code{\link{wq_edsm}}), "DJFMP" (Delta Juvenile Fish Monitoring Program, \code{\link{wq_djfmp}}), "20mm" (20mm Survey, \code{\link{wq_20mm}}), "SKT" (Spring Kodiak Trawl, \code{\link{wq_skt}}), "Baystudy" (Bay Study, \code{\link{wq_baystudy}}), "USBR" (United States Bureau of Reclamation Sacramento Deepwater Ship Channel data, \code{\link{wq_usbr}}), and "Suisun" (Suisun Marsh Fish Study, \code{\link{wq_suisun}}).}

\item{Shapefile}{Shapefile you would like used to define regions in the dataset. Must be in \code{\link[sf]{sf}} format, e.g., imported with \code{\link[sf]{st_read}}. Defaults to \code{\link{deltaregions}}.}

\item{Region_column}{Quoted name of the column in the Shapefile with the region designations.}

\item{Regions}{Character vector of regions to be included in the dataset. Must correspond with levels of the \code{Region_column}. To include all data points regardless of whether they correspond to a region in the \code{Shapefile} set \code{Regions = NULL}.}
}
\value{
A list of datasets
}
\description{
Imports, filters, and processes datasets and outputs a list of desired datasets
}
\examples{
Data <- DeltaDater(Start_year = 1900,
WQ_sources = c("EMP", "STN", "FMWT", "EDSM", "DJFMP", "SKT", "20mm", "Suisun", "Baystudy", "USBR"),
Variables = "Water quality",
Regions = NULL)
}
