% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaDayflower.R
\name{DeltaDayFlower}
\alias{DeltaDayFlower}
\title{Plot dayflow data}
\usage{
DeltaDayFlower(Start_year = 2002, End_year = 2018)
}
\arguments{
\item{Start_year}{First year (integer) that should be included in the plot}

\item{End_year}{Last year (integer) that should be included in the plot. This year will also be highlighted.}
}
\value{
A list with the plot and processed data.
}
\description{
Function to process and plot dayflow data
}
\seealso{
\code{\link{DeltaDater}}
}
