% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaWQer.R
\name{DeltaWQer}
\alias{DeltaWQer}
\title{Plot water quality data}
\usage{
DeltaWQer(
  Data,
  Start_year = 2002,
  End_year = 2018,
  Regions = c("Suisun Bay", "Suisun Marsh", "Lower Sacramento River",
    "Sac Deep Water Shipping Channel", "Cache Slough/Liberty Island",
    "Lower Joaquin River", "Southern Delta"),
  Temp_season = "Summer",
  Secchi_season = "Fall",
  Salinity_season = "Fall",
  Chl_season = "Summer",
  Micro_season = "Summer"
)
}
\arguments{
\item{Data}{Input dataset created by \code{\link{DeltaDater}}.}

\item{Start_year}{First year (integer) that should be included in the plot}

\item{End_year}{Last year (integer) that should be included in the plot. This year will also be highlighted.}

\item{Regions}{Character vector of regions to include in the plot. The data will be filtered to only include these regions and ordered in the order provided here. To include data with NA regions, set \code{Regions=NULL}.}

\item{Temp_season}{Character vector of seasons to retain for temperature. Should be one of "Summer", "Fall", "Winter", or "Spring".}

\item{Secchi_season}{Character vector of seasons to retain for Secchi. Should be one of "Summer", "Fall", "Winter", or "Spring".}

\item{Salinity_season}{Character vector of seasons to retain for salinity. Should be one of "Summer", "Fall", "Winter", or "Spring".}

\item{Chl_season}{Character vector of seasons to retain for chlorophyll. Should be one of "Summer", "Fall", "Winter", or "Spring".}

\item{Micro_season}{Character vector of seasons to retain for \emph{Microcystis}. Should be one of "Summer", "Fall", "Winter", or "Spring".}
}
\value{
A list with the plot and processed data.
}
\description{
Function to process and plot water quality (temperature, Secchi depth, salinity, chlorophyll, and \emph{Microcystis}) data
}
\seealso{
\code{\link{DeltaDater}}
}
