% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_edsm}
\alias{wq_edsm}
\title{EDSM water quality data}
\format{
a tibble with 14,851 rows and 12 variables
\describe{
  \item{Source}{Name of the source dataset.}
  \item{Station}{Station where sample was collected.}
  \item{Latitude}{Latitude in decimal degrees.}
  \item{Longitude}{Longitude in decimal degrees.}
  \item{Date}{Date sample was collected.}
  \item{Datetime}{Date and time of sample collection.}
  \item{Depth}{Bottom depth (m).}
  \item{Tide}{Tidal stage.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Temperature}{Temperature in °C.}
  \item{Temperature_bottom}{Temperature (°C) at bottom.}
  \item{Notes}{Comments.}
  }
}
\usage{
wq_edsm
}
\description{
Water quality data from the United States Fish and Wildlife Service Enhanced Delta Smelt Monitoring Program.
}
\details{
More metadata and information on methods are available \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.415.1}{here}.
}
\seealso{
\code{\link{DeltaDater}}, \code{\link{DeltaWQer}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
