% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_emp}
\alias{wq_emp}
\title{EMP water quality data}
\format{
a tibble with 16,293 rows and 11 variables
\describe{
  \item{Source}{Name of the source dataset.}
  \item{Station}{Station where sample was collected.}
  \item{Date}{Date sample was collected.}
  \item{Datetime}{Date and time sample was collected.}
  \item{Tide}{Tidal stage (always High Slack).}
  \item{Microcystis}{Microcystis bloom intensity on a qualitative scale from 1 to 5 where 1 = absent, 2 = low, 3 = medium, 4 = high, and 5 = very high.}
  \item{Chlorophyll}{Chlorophyll concentration (\eqn{\mu}g \ifelse{html}{\out{L<sup>-1</sup>}}{\eqn{L^{-1}}}).}
  \item{Secchi}{Secchi depth (cm).}
  \item{Temperature}{Temperature (°C) at surface.}
  \item{Temperature_bottom}{Temperature (°C) at bottom.}
  \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
  \item{Notes}{Notes or comments.}
  }
}
\usage{
wq_emp
}
\description{
Water quality data from the California Department of Water Resources Environmental Monitoring Program.
}
\details{
More metadata and information on methods are available \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.458.2}{here}.
}
\seealso{
\code{\link{DeltaDater}}, \code{\link{DeltaWQer}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
