% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wq_stn}
\alias{wq_stn}
\title{STN water quality data}
\format{
a tibble with 8,266 rows and 12 variables
\describe{
  \item{Source}{Name of the source dataset.}
  \item{Station}{Station where sample was collected.}
  \item{Date}{Date sample was collected.}
  \item{Datetime}{Date and time of sample collection.}
  \item{Depth}{Bottom depth (m).}
  \item{Tide}{Tidal stage.}
  \item{Microcystis}{Microcystis bloom intensity on a qualitative scale from 1 to 5 where 1 = absent, 2 = low, 3 = medium, 4 = high, and 5 = very high.}
  \item{Secchi}{Secchi depth (cm).}
  \item{Temperature}{Temperature (°C) at surface.}
  \item{Temperature_bottom}{Temperature (°C) at bottom.}
  \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
  \item{Notes}{Comments.}
  }
}
\usage{
wq_stn
}
\description{
Water quality data from the California Department of Fish and Wildlife Summer Townet Survey.
}
\details{
More metadata and information on methods are available \href{https://wildlife.ca.gov/Conservation/Delta/Townet-Survey}{here}.
}
\seealso{
\code{\link{DeltaDater}}, \code{\link{DeltaWQer}}, \code{\link{DeltaMetadater}}
}
\keyword{datasets}
